cifar10_diffusion/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
tqdm
numpy
Pillow
modal>=0.64.0
